// src/services/rest-api/customer-service/model/spouse-info-model.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Customer Service API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.242
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface SpouseInfoModel
 */
export interface SpouseInfoModel {
    /**
     * 
     * @type {string}
     * @memberof SpouseInfoModel
     */
    'familyStatus'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof SpouseInfoModel
     */
    'titleCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof SpouseInfoModel
     */
    'titleName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof SpouseInfoModel
     */
    'firstName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof SpouseInfoModel
     */
    'lastName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof SpouseInfoModel
     */
    'jobWorkPlace'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof SpouseInfoModel
     */
    'jobTypeCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof SpouseInfoModel
     */
    'jobPosition'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof SpouseInfoModel
     */
    'jobYoS'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof SpouseInfoModel
     */
    'income'?: number;
    /**
     * 
     * @type {number}
     * @memberof SpouseInfoModel
     */
    'incomeOther'?: number;
    /**
     * 
     * @type {string}
     * @memberof SpouseInfoModel
     */
    'incomeSourceOther'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof SpouseInfoModel
     */
    'incomeSourceOversea'?: string | null;
}